added tests and proper cost deduction. Also moved from Float cash to Numeric10,2 cash

+++ app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, Boolean
+from sqlalchemy import Column, Integer, Numeric, String, DateTime, Boolean
 from datetime import datetime, timezone
 
 from .base import Base
@@ -8,7 +8,7 @@ class User(Base):
     
     uid = Column(Integer, primary_key=True, index=True)
     name = Column(String(255))
-    cash = Column(Float)
+    cash = Column(Numeric(10, 2))
     creation_time = Column(DateTime, default=datetime.now(timezone.utc))
     hashed_password = Column(String(255))
     is_admin = Column(Boolean, default=False)
@@ -17,7 +17,7 @@ class User(Base):
         return {
             "uid": self.uid,
             "name": self.name,
-            "cash": self.cash,
+            "cash": float(self.cash),
             "creation_time": self.creation_time,
             "is_admin": self.is_admin
         }
